Refactor i18n utilities for test compatibility
